requirements.txt
colorama
hwi
pyside2
shiboken2
structlog
python-bitcoinrpc
